+++ packages/styled/index.d.ts
@@ -4,10 +4,15 @@ type StyledComponent = <T>(
   strings: TemplateStringsArray,
   ...args: StyledComponentArg<T>[]
 ) => React.FC<Props>;
+type GeneralStyledComponent<T> = (
+  strings: TemplateStringsArray,
+  ...args: StyledComponentArg<T>[]
+) => React.FC<T>;
 
 interface IStyled {
   [name: string]: StyledComponent;
 }
-declare const styled: IStyled;
+declare const styled: IStyled &
+  (<T>(Base: React.ReactNode) => GeneralStyledComponent<T>);
 
 export = styled;
